requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
markdown
django-filter
django-cors-headers
graphene_django
pytest-django==4.4.0
model-bakery==1.3.2
factory-boy==3.2.0
gunicorn==20.1.0
django-storages
boto3